# openapi/specs/openapi.yaml
openapi: 3.1.0
info:
  title: GoPlace
  version: 0.0.1
servers:
  - url: "https://api-dev.goplace.io"
    description: Development server
  - url: "https://api-staging.goplace.io"
    description: Staging server
  - url: "https://api.goplace.io"
    description: Production server
...
paths:
...
  /reservations/statuses:
    get:
      summary: Get all reservation statuses
      description: Get all reservation statuses
      operationId: getAllReservationStatuses
      parameters:
        - name: shortData
          in: query
          description: Whether to get short response or full
          required: false
          schema:
            type: boolean
        - name: branchIds
          in: query
          description: Branch ids
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Reservation status body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllReservationStatusesResponse"
        default:
          $ref: "#/components/responses/GeneralError"
    post:
      summary: Create a reservation status
      description: Create a reservation status
      operationId: createReservationStatus
      requestBody:
        description: Reservation status body
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationStatusRequest"
      responses:
        "200":
          description: Body of created status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateReservationStatusResponse"
        default:
          $ref: "#/components/responses/GeneralError"
...
  /shifts/{shiftId}:
    get:
      summary: Get shift by id
      description: Get shift by id
      operationId: getShiftByID
      parameters:
        - name: shiftId
          in: path
          description: Auto incremental shift id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetShiftByIDResponse"
        default:
          $ref: "#/components/responses/GeneralError"
    put:
      summary: Update shift data
      description: Update shift data
      operationId: updateShift
      x-permissions: ["update.shifts"]
      parameters:
        - name: shiftId
          in: path
          description: Auto incremental shift id
          required: true
          schema:
            type: integer
            format: int32
        - name: editMode
          in: query
          description: The mode to edit the shift (current, following or all)
          required: false
          schema:
            type: string
        - name: deleteMode
          in: query
          description: The mode to delete the shift (current, following or all)
          required: false
          schema:
            type: string
        - name: date
          in: query
          description: The day to edit
          required: false
          schema:
            type: string
      requestBody:
        description: Shift Id and new data
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShiftRequest"
      responses:
        "201":
          description: Shift updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateShiftResponse"
        default:
          $ref: "#/components/responses/GeneralError"
...
  /users/roles:
    get:
      summary: Get all roles
      description: Get all roles
      operationId: getAllRoles
      x-permissions: ["access.users"]
      parameters:
        - name: perPage
          in: query
          description: Number of records per page
          schema:
            type: integer
            format: int32
        - name: currentPage
          in: query
          description: Page number
          schema:
            type: integer
            format: int32
        - name: department
          in: query
          description: Department ids
          schema:
            type: integer
            format: int32
        - name: query
          in: query
          description: Query
          schema:
            type: string
      responses:
        "200":
          description: All roles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllRolesResponse"
        default:
          $ref: "#/components/responses/GeneralError"
    post:
      summary: Create new role
      description: Create new role
      x-permissions: ["access.users"]
      operationId: createRole
      requestBody:
        description: Role body
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "200":
          description: Created role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRoleResponse"
        default:
          $ref: "#/components/responses/GeneralError"
...
  /restaurant-items/{itemId}:
    get:
      summary: Get restaurant item
      description: Get restaurant item
      operationId: getRestaurantItemByID
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRestaurantItemByIDResponse"
        default:
          $ref: "#/components/responses/GeneralError"
    put:
      summary: Update restaurant item
      description: Update restaurant item
      operationId: updateRestaurantItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRestaurantItemRequest"
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateRestaurantItemResponse"
        default:
          $ref: "#/components/responses/GeneralError"
    delete:
      summary: Delete restaurant item
      description: Delete restaurant item
      operationId: deleteRestaurantItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteRestaurantItemResponse"
        default:
          $ref: "#/components/responses/GeneralError"
...
components:
  responses:
    GeneralError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GeneralError"
  schemas:
...
    GeneralError:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
        - code
        - message

    # Login
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          format: int32
        perPage:
          type: integer
          format: int32
        currentPage:
          type: integer
          format: int32
        lastPage:
          type: integer
          format: int32
        from:
          type: integer
          format: int32
        to:
          type: integer
          format: int32
      required:
        - total
        - perPage
        - currentPage
        - lastPage
        - from
        - to

    # Guest
...
    ReservationStatus:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        color:
          type: string
        category:
          type: string
        type:
          type: string
        icon:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - name
        - color
        - icon
        - category
        - type
    ReservationStatusParams:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        category:
          type: string
        type:
          type: string
        color:
          type: string
        icon:
          type: string
      required:
        - name
...
    Shift:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        from:
          type: string
        to:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        timeInterval:
          type: integer
          format: int32
        floorPlan:
          $ref: "#/components/schemas/ShortFloorPlan"
        seatingAreas:
          type: array
          items:
            $ref: "#/components/schemas/SeatingArea"
        category:
          $ref: "#/components/schemas/ShiftCategory"
        minGuests:
          type: integer
          format: int32
        maxGuests:
          type: integer
          format: int32
        daysToRepeat:
          type: array
          items:
            type: string
        turnover:
          type: array
          items:
            $ref: "#/components/schemas/Turnover"
        pacing:
          type: array
          items:
            $ref: "#/components/schemas/Pacing"
        exceptions:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - name
        - from
        - to
        - startDate
        - endDate
        - timeInterval
        - floorPlan
        - seatingArea
        - category
        - minGuests
        - maxGuests
        - daysToRepeat
        - seatingAreas
        - turnover
        - exceptions
        - pacing
    ShiftParams:
      type: object
      properties:
        id:
          type: integer
          format: int32
        branchId:
          type: integer
          format: int32
        name:
          type: string
        from:
          type: string
        to:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        timeInterval:
          type: integer
          format: int32
        floorPlanId:
          type: integer
          format: int32
        seatingAreaIds:
          type: array
          items:
            type: integer
            format: int32
        categoryId:
          type: integer
          format: int32
        minGuests:
          type: integer
          format: int32
        maxGuests:
          type: integer
          format: int32
        daysToRepeat:
          type: array
          items:
            type: string
        turnover:
          type: array
          items:
            $ref: "#/components/schemas/TurnoverParams"
        pacing:
          type: array
          items:
            $ref: "#/components/schemas/PacingParams"
    Pacing:
      type: object
      properties:
        id:
          type: integer
          format: int32
        shiftId:
          type: integer
          format: int32
        seatingAreaId:
          type: integer
          format: int32
        hour:
          type: string
        capacity:
          type: integer
          format: int32
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - shiftId
        - seatingAreaId
        - hour
        - capacity
    PacingParams:
      type: object
      properties:
        seatingAreaId:
          type: integer
          format: int32
        hour:
          type: string
        capacity:
          type: integer
          format: int32
    ShiftCategory:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ShortFloorPlan:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
      required:
        - name
...
    Turnover:
      type: object
      properties:
        id:
          type: integer
          format: int32
        shiftId:
          type: integer
          format: int32
        guestsNumber:
          type: integer
          format: int32
        turnoverTime:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    TurnoverParams:
      type: object
      properties:
        id:
          type: integer
          format: int32
        shiftId:
          type: integer
          format: int32
        guestsNumber:
          type: integer
          format: int32
        turnoverTime:
          type: string
    SeatingArea:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        branchId:
          type: integer
          format: int32
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - name
...
    CreateReservationStatusRequest:
      $ref: "#/components/schemas/ReservationStatusParams"
    CreateReservationStatusResponse:
      $ref: "#/components/schemas/ReservationStatus"
    BranchReservationStatus:
      type: object
      properties:
        branch:
          type: string
        statuses:
          type: array
          items:
            $ref: "#/components/schemas/ReservationStatus"
    GetAllReservationStatusesResponse:
      type: array
      items:
        $ref: "#/components/schemas/BranchReservationStatus"
...
    GetShiftByIDResponse:
      $ref: "#/components/schemas/Shift"
    UpdateShiftRequest:
      $ref: "#/components/schemas/ShiftParams"
    UpdateShiftResponse:
      $ref: "#/components/schemas/Shift"
...
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int32
        department:
          type: string
        name:
          type: string
        displayName:
          type: string
        permissions:
          type: array
          items:
            type: string
        usersCount:
          type: integer
          format: int32
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - department
        - name
        - displayName
        - permissions
        - usersCount
        - createdAt
        - updatedAt
    RoleParams:
      type: object
      properties:
        id:
          type: integer
          format: int32
        department:
          type: integer
          format: int32
        permissions:
          type: array
          nullable: true
          items:
            type: integer
            format: int32
        name:
          type: string
        displayName:
          type: string
...
    CreateRoleRequest:
      $ref: "#/components/schemas/RoleParams"
    CreateRoleResponse:
      $ref: "#/components/schemas/Role"
...
    GetAllRolesResponse:
      type: object
      properties:
        pagination:
          $ref: "#/components/schemas/Pagination"
        result:
          type: array
          items:
            $ref: "#/components/schemas/Role"
      required:
        - pagination
        - result
...
    RestaurantItemParams:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        price:
          type: number
          format: float
        image:
          type: string
        icon:
          type: string
    RestaurantItem:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        price:
          type: number
          format: float
        code:
          type: string
        image:
          type: string
        icon:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    UpdateRestaurantItemRequest:
      $ref: "#/components/schemas/RestaurantItemParams"
    UpdateRestaurantItemResponse:
      $ref: "#/components/schemas/RestaurantItem"
...
    GetRestaurantItemByIDResponse:
      $ref: "#/components/schemas/RestaurantItem"
...
    DeleteRestaurantItemResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT